requirements.txt
pendulum==2.1.2
pandas==1.5.3
geopandas==0.12.2
rasterio==1.3.6
scipy==1.10.1
rasterstats==0.18.0
matplotlib==3.7.1
psycopg2==2.9.5
python-dotenv==1.0.0